moved functions to dataframe functions
