+++ dashboard/pages/3_Sentiment_Over_Time.py
@@ -9,23 +9,13 @@ import pandas as pd
 
 from db_functions import get_scores_topic, get_topic_names
 from d_graphs import visualise_change_over_time, visualise_heatmap
+from dataframe_functions import resample_dataframe, add_year_month_day_columns
 
-AGGREGATES = ["mean", "count"]
 
-
-@st.cache_data
-def resample_dataframe(df: pd.DataFrame, time_interval: str, aggregate: str):
-    """Resamples the dataframe to return the aggregate sentiment scores by 
-    (source, topic) over a set of grouped time intervals."""
-    if not aggregate in AGGREGATES:
-        raise ValueError(
-            f"The aggregate parameter must be one of {AGGREGATES}.")
-
-    df_avg = df.groupby(['source_name', 'topic_name']).resample(
-        time_interval, on='date_published').agg({"title_polarity_score": aggregate,
-                                                 "content_polarity_score": aggregate}).reset_index()
-
-    return pd.DataFrame(df_avg)
+def select_data_by_topic(topic_name: str) -> pd.DataFrame:
+    """Selects the data for a topic by name"""
+    data = get_scores_topic(topic_name)
+    return pd.DataFrame(data)
 
 
 def construct_streamlit_time_graph(data_df: pd.DataFrame, avg_col: DeltaGenerator,
@@ -48,20 +38,6 @@ def construct_streamlit_time_graph(data_df: pd.DataFrame, avg_col: DeltaGenerato
     count_col.altair_chart(count_graph, use_container_width=True)
 
 
-def add_year_month_day_columns(data_df: pd.DataFrame) -> pd.DataFrame:
-    """Adds year, week, and weekday columns to a dataframe"""
-    data_df["year"] = data_df["date_published"].dt.year
-    data_df["week_num"] = data_df["date_published"].dt.isocalendar().week
-    data_df["month_name"] = data_df["date_published"].dt.strftime('%b')
-    data_df["week_of_month"] = data_df["date_published"].apply(
-        lambda d: (d.day - 1) // 7 + 1)
-    data_df["week_text"] = data_df["month_name"] + \
-        " Week " + data_df["week_of_month"].astype(str)
-    data_df["weekday"] = data_df["date_published"].dt.day_name()
-    data_df["date_name"] = data_df["date_published"].dt.strftime('%d-%m-%Y')
-    return data_df
-
-
 def construct_streamlit_heatmap(heatmaps_container: DeltaGenerator,
                                 weekly_data: pd.DataFrame, sent_by_title: bool,
                                 colour: str = 'yellowgreen'):
@@ -82,12 +58,6 @@ def construct_sidebar(topics_list: list[str]) -> tuple[str, str]:
     granularity = st.sidebar.selectbox(
         "Granularity", granularity_to_hours.keys())
     return topic, granularity_to_hours[granularity]
-
-
-def select_data_by_topic(topic_name: str) -> pd.DataFrame:
-    """Selects the data for a topic by name"""
-    data = get_scores_topic(topic_name)
-    return pd.DataFrame(data)
 
 
 def construct_linegraphs_container() -> list[list]:
